+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Add option to use a CDN
 - Add option for max file size
+- Add option to rotate image
 
 ### improvements