+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## 1.0.1 (unreleased)
+## 1.1.0 (March 30, 2014)
 
 ### features